requirements.txt
transformers>=4.41.0
sentencepiece>=0.2.0
pypdf>=5.0.0
streamlit>=1.37.0
accelerate>=0.33.0
huggingface-hub>=0.24.0
numpy>=1.26.0
torch>=2.0.0

python-telegram-bot>=20.7
requests>=2.32.0
